+++ frontend/src/components/Modal/IntroModal/IntroModal.css
@@ -1,0 +1,17 @@
+.buttons {
+    margin-left: 43%;
+    
+    display: flex;
+    justify-content: flex-start;
+    flex-flow: row wrap;
+    align-items: flex-start;
+}
+
+.signup-open-button {
+    display: flex;
+    margin-right: 2%;
+}
+
+.signin-open-button {
+    display: flex;
+}
