+++ frontend/src/components/index.js
@@ -4,8 +4,7 @@ import PaperSpec from "./Paper/PaperSpec/PaperSpec";
 import PaperCard from "./Paper/PaperCard/PaperCard";
 import ReviewCard from "./Review/ReviewCard/ReviewCard";
 import Header from "./Header/Header";
-import Signup from "./Modal/Signup/Signup";
-import Signin from "./Modal/Signin/Signin";
+import IntroModal from "./Modal/IntroModal/IntroModal";
 import ReviewReply from "./Reply/ReviewReply/ReviewReply";
 
 export {
@@ -15,7 +14,6 @@ export {
     PaperCard,
     ReviewCard,
     Header,
-    Signup,
-    Signin,
+    IntroModal,
     ReviewReply,
 };
